+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Adjusted JSON schema to only require the very essential defaults.
 * Added trimming to selectedText, clipboardContext, appContext.
 * New repo for [Alfred Workflow](https://github.com/ognistik/macrowhisper/tree/main/alfred)!
+* Improvements for URL encoding in actions
 
 ## [v1.2.3](https://github.com/ognistik/macrowhisper/releases/tag/v1.2.3) - 2025/08/06
 ### Added